requirements.txt
python-kasa
aioping
